move context-menu to the correct folder
