+++ src/app/components/context-menu/animations.ts
