+++ src/app/components/context-menu/context-menu.component.ts
@@ -31,15 +31,15 @@ import { trigger } from '@angular/animations';
 import { FocusKeyManager } from '@angular/cdk/a11y';
 import { DOWN_ARROW, UP_ARROW } from '@angular/cdk/keycodes';
 
-import { ExtensionService } from '../extensions/extension.service';
-import { AppStore, SelectionState } from '../store/states';
-import { appSelection } from '../store/selectors/app.selectors';
+import { ExtensionService } from '../../extensions/extension.service';
+import { AppStore, SelectionState } from '../../store/states';
+import { appSelection } from '../../store/selectors/app.selectors';
 import { Store } from '@ngrx/store';
 import { Subject } from 'rxjs/Rx';
 import { takeUntil } from 'rxjs/operators';
 
 import { ContextMenuOverlayRef } from './context-menu-overlay';
-import { ContentActionRef } from '../extensions/action.extensions';
+import { ContentActionRef } from '../../extensions/action.extensions';
 import { contextMenuAnimation } from './animations';
 import { ContextMenuItemDirective } from './context-menu-item.directive';
