+++ src/app/components/context-menu/interfaces.ts
