+++ src/app/ui/custom-theme.scss
@@ -8,8 +8,8 @@
 @import '../components/settings/settings.component.theme';
 @import '../components/current-user/current-user.component.theme';
 @import '../components/permission-manager/permissions-manager.component.theme';
+@import '../components/context-menu/context-menu.component.theme';
 @import '../dialogs/node-versions/node-versions.dialog.theme';
-@import '../context-menu/context-menu.component.theme';
 
 @import './overrides/adf-toolbar.theme';
 @import './overrides/adf-search-filter.theme';
